Task1_RAG_Chabot/requirements.txt
streamlit
PyPDF2
faiss-cpu
sentence-transformers
numpy
nest_asyncio
